move resampling code to separate folder
