+++ src/openms/include/OpenMS/FILTERING/RESAMPLING/sources.cmake
@@ -1,15 +1,10 @@
 ### the directory name
-set(directory include/OpenMS/FILTERING/TRANSFORMERS)
+set(directory include/OpenMS/FILTERING/RESAMPLING)
 
 ### list all header files of the directory here
 set(sources_list_h
-NLargest.h
-Normalizer.h
-RankScaler.h
-SpectraMerger.h
-SqrtScaler.h
-ThresholdMower.h
-WindowMower.h
+LinearResampler.h
+LinearResamplerAlign.h
 )
 
 ### add path to the filenames
@@ -19,7 +14,7 @@ foreach(i ${sources_list_h})
 endforeach(i)
 
 ### source group definition
-source_group("Header Files\\OpenMS\\FILTERING\\TRANSFORMERS" FILES ${sources_h})
+source_group("Header Files\\OpenMS\\FILTERING\\RESAMPLING" FILES ${sources_h})
 
 set(OpenMS_sources_h ${OpenMS_sources_h} ${sources_h})
