+++ src/openms/includes.cmake
@@ -42,6 +42,7 @@ include(source/ANALYSIS/TOPDOWN/sources.cmake)
 include(source/ANALYSIS/XLMS/sources.cmake)
 include(source/FILTERING/BASELINE/sources.cmake)
 include(source/FILTERING/TRANSFORMERS/sources.cmake)
+include(source/FILTERING/RESAMPLING/sources.cmake)
 include(source/FILTERING/DATAREDUCTION/sources.cmake)
 include(source/FILTERING/CALIBRATION/sources.cmake)
 include(source/FILTERING/SMOOTHING/sources.cmake)
@@ -109,6 +110,7 @@ include(include/OpenMS/FILTERING/ID/sources.cmake)
 include(include/OpenMS/FILTERING/NOISEESTIMATION/sources.cmake)
 include(include/OpenMS/FILTERING/SMOOTHING/sources.cmake)
 include(include/OpenMS/FILTERING/TRANSFORMERS/sources.cmake)
+include(include/OpenMS/FILTERING/RESAMPLING/sources.cmake)
 include(include/OpenMS/FEATUREFINDER/sources.cmake)
 include(include/OpenMS/FILTERING/CENTROIDING/sources.cmake)
 include(include/OpenMS/QC/sources.cmake)
