+++ src/openms/source/FEATUREFINDER/FeatureFinderAlgorithmMRM.cpp
@@ -11,7 +11,7 @@
 #include <OpenMS/FEATUREFINDER/EmgModel.h>
 #include <OpenMS/FILTERING/SMOOTHING/GaussFilter.h>
 #include <OpenMS/FILTERING/NOISEESTIMATION/SignalToNoiseEstimatorMeanIterative.h>
-#include <OpenMS/FILTERING/TRANSFORMERS/LinearResampler.h>
+#include <OpenMS/FILTERING/RESAMPLING/LinearResampler.h>
 
 #include <fstream>
 #include <map>
