+++ src/tests/class_tests/openms/source/LinearResampler_test.cpp
@@ -10,7 +10,7 @@
 #include <OpenMS/test_config.h>
 
 
-#include <OpenMS/FILTERING/TRANSFORMERS/LinearResampler.h>
+#include <OpenMS/FILTERING/RESAMPLING/LinearResampler.h>
 
 #include <OpenMS/KERNEL/MSExperiment.h>
 #include <OpenMS/KERNEL/StandardTypes.h>
